Test builtin-stringop-chk-1.c now passes on or1k.
One of the tests from builtin-stringop-chk-1.c is expected to produce an
error message.  Previously this test failed for or1k, the error message
was not produced.

At some point the test started working, the error message is now
produced.

Commit ddb9f9e4 modified the test file so that the error message was not
expected for or1k.  This commit reverts the part of ddb9f9e4 for this
test file so that the error message is once again expected.

gcc/testsuite/ChangeLog:

	* gcc.dg/builtin-stringop-chk-1.c: Revert commit ddb9f9e4 on this
	file, this test passes on or1k now.